Add files via upload
Paragraph segmentation is the process of dividing the meaningful text into categories in which they belong. It is also a very important concept and used widely in machine learning. 

The code consists of two very important concepts. one if TFIDF and the other is Cosine Similarity (also used in widely C99 algorithm) to group text together.